# Set optimizeForLatency in decoder configuration