--- a/samples/webcodecs-echo/js/stream_worker.js
+++ b/samples/webcodecs-echo/js/stream_worker.js
@@ -719,6 +719,8 @@ SSRC = this.config.ssrc
              if (cfg.decoderConfig) {
                // self.postMessage({text: 'Decoder reconfig!'});
                cfg.decoderConfig.hardwareAcceleration = config.decHwAcceleration;
+               cfg.decoderConfig.optimizeForLatency = true;
+               if (config.latencyPref == 'quality') cfg.decoderConfig.optimizeForLatency = false;
                const decoderConfig = JSON.stringify(cfg.decoderConfig);
                // self.postMessage({text: 'Decoder configuration: ' + decoderConfig});
                const configChunk =