--- a/samples/webcodecs-echo/js/stream_worker.js
+++ b/samples/webcodecs-echo/js/stream_worker.js
@@ -719,7 +719,7 @@ SSRC = this.config.ssrc
              if (cfg.decoderConfig) {
                // self.postMessage({text: 'Decoder reconfig!'});
                cfg.decoderConfig.hardwareAcceleration = config.decHwAcceleration;
-               cfg.decoderConfig.optimizeForLatency = !(config.latencyPref == 'quality');
+               cfg.decoderConfig.optimizeForLatency = (config.latencyPref == "realtime");
                const decoderConfig = JSON.stringify(cfg.decoderConfig);
                // self.postMessage({text: 'Decoder configuration: ' + decoderConfig});
                const configChunk =